prediction-service/pyproject.toml
[project]
name = "prediction-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.134.0",
    "mistralai>=1.12.4",
    "python-dotenv>=1.2.1",
    "uvicorn[standard]>=0.41.0",
]
